block id stored in a short blocks pos now relative to the chunk

+++ MineTest/res/shader/BlockRenderer.vert
@@ -4,10 +4,10 @@ layout (location = 0) in vec3 block_vertex;
 layout (location = 1) in vec2 block_texture_coord;
 
 layout (location = 2) in vec3 position;
-layout (location = 3) in float id;
+layout (location = 3) in int id;
 
 out vec2 TexCoords;
-out float ID;
+out int ID;
 
 uniform mat4 vp;
